src/third_party/blink/web_tests/external/wpt/css/css-text/parsing/hyphens-computed-expected.txt

This is a testharness.js-based test.
PASS Property hyphens value 'none' computes to 'none'
PASS Property hyphens value 'manual' computes to 'manual'
FAIL Property hyphens value 'auto' computes to 'auto' assert_equals: expected "auto" but got "manual"
Harness: the test ran to completion.
